Remove ModelSearchTestCase
This provides no additional functionality over TestCase, and was introduced as a replacement for WagtailTestUtils, which none of the tests that inherited from it made use of.

In particular, modelsearch.tests.test_backends.BackendTests should not subclass any TestCase class - it should act as a mixin for the backend-specific test classes, which may specify TestCase or TransactionTestCase as appropriate. TestMySQLSearchBackend only works as a TransactionTestCase, hence why MySQL tests were failing.