Reinstate MySQL / MariaDB support (#19)
* Add vagrant environment from https://github.com/wagtail/wagtailsearch/

* Allow passing ca_certs from the environment

* Add task for running coverage report within vagrant

* Add Opensearch 1 to Vagrant setup

* Fill in the correct architecture for opensearch downloads

* Add Elasticsearch 9 to Vagrant setup

* Add documentation for running code coverage reports under Vagrant

* Revert "Drop support for MySQL/MariaDB"

This reverts commit 3d610119f909d534e56d1091b9792c337bd501c4.

* Remove ngram parser from autocomplete index on MySQL

As per https://github.com/wagtail/wagtail/commit/0a7353ef78aa5e18a23528a75b1bef7830009cd2

* Remove ModelSearchTestCase

This provides no additional functionality over TestCase, and was introduced as a replacement for WagtailTestUtils, which none of the tests that inherited from it made use of.

In particular, modelsearch.tests.test_backends.BackendTests should not subclass any TestCase class - it should act as a mixin for the backend-specific test classes, which may specify TestCase or TransactionTestCase as appropriate. TestMySQLSearchBackend only works as a TransactionTestCase, hence why MySQL tests were failing.

* Add some more missing mentions of MySQL and MariaDB to the docs

* Fix db backend path in test settings

This isn't used by the tests anyway, because the database backend test classes override MODELSEARCH_BACKENDS for themselves, but we might as well be correct...

---------

Co-authored-by: Karl Hobley <karl@kaed.uk>